HADOOP-14914. Change to a safely casting long to int. Contributed by Ajay Kumar.

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-nodemanager/src/main/java/org/apache/hadoop/yarn/server/nodemanager/DirectoryCollection.java
@@ -605,7 +605,7 @@ public class DirectoryCollection {
     if (totalSpace != 0) {
       long tmp = ((totalSpace - usableSpace) * 100) / totalSpace;
       if (Integer.MIN_VALUE < tmp && Integer.MAX_VALUE > tmp) {
-        goodDirsDiskUtilizationPercentage = (int) tmp;
+        goodDirsDiskUtilizationPercentage = Math.toIntExact(tmp);
       }
     } else {
       // got no good dirs
